refactor: extract destructuring transform routines
